# backend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kangur Platform API",
    "version": "0.3.0",
    "description": "AUTO-GENERATED from Zod via openapi:generate. Do not hand-edit."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "AUTH_REQUIRED",
              "INVALID_TOKEN",
              "TOKEN_EXPIRED",
              "NOT_FOUND",
              "VALIDATION_ERROR",
              "FORBIDDEN",
              "CONFLICT",
              "INSUFFICIENT_CREDITS"
            ]
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": {}
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "NotificationPreferences": {
        "type": "object",
        "properties": {
          "silentMode": {
            "type": "boolean"
          },
          "workspaceInvitations": {
            "type": "boolean"
          },
          "shoppingStarted": {
            "type": "boolean"
          },
          "shoppingFinished": {
            "type": "boolean"
          },
          "shoppingListCreated": {
            "type": "boolean"
          },
          "shoppingListDeleted": {
            "type": "boolean"
          },
          "appUpdates": {
            "type": "boolean"
          },
          "newFeatures": {
            "type": "boolean"
          },
          "offersPromos": {
            "type": "boolean"
          }
        },
        "required": [
          "silentMode",
          "workspaceInvitations",
          "shoppingStarted",
          "shoppingFinished",
          "shoppingListCreated",
          "shoppingListDeleted",
          "appUpdates",
          "newFeatures",
          "offersPromos"
        ]
      },
      "UpdateNotificationPreferencesBody": {
        "type": "object",
        "properties": {
          "silentMode": {
            "type": "boolean"
          },
          "workspaceInvitations": {
            "type": "boolean"
          },
          "shoppingStarted": {
            "type": "boolean"
          },
          "shoppingFinished": {
            "type": "boolean"
          },
          "shoppingListCreated": {
            "type": "boolean"
          },
          "shoppingListDeleted": {
            "type": "boolean"
          },
          "appUpdates": {
            "type": "boolean"
          },
          "newFeatures": {
            "type": "boolean"
          },
          "offersPromos": {
            "type": "boolean"
          }
        }
      },
      "RegisterPushDeviceBody": {
        "type": "object",
        "properties": {
          "expoToken": {
            "type": "string",
            "minLength": 1
          },
          "platform": {
            "type": "string",
            "enum": [
              "ios",
              "android",
              "web"
            ]
          },
          "appVersion": {
            "type": "string"
          }
        },
        "required": [
          "expoToken"
        ]
      },
      "UnregisterPushDeviceBody": {
        "type": "object",
        "properties": {
          "expoToken": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "expoToken"
        ]
      },
...
      "ShoppingListDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "workspaceId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "emoji": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ShoppingListStatus"
          },
          "isUntitled": {
            "type": "boolean"
          },
          "itemCount": {
            "type": "integer",
            "minimum": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "workspaceId",
          "name",
          "emoji",
          "status",
          "isUntitled",
          "itemCount",
          "createdAt",
          "updatedAt"
        ]
      },
      "ShoppingListStatus": {
        "type": "string",
        "enum": [
          "active",
          "archived"
        ]
      },
...
      "UpdateShoppingListBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "emoji": {
            "type": "string"
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/workspaces/{workspaceId}/invitations/{invitationId}": {
      "delete": {
        "summary": "Revoke a pending invitation",
        "tags": [
          "Workspaces"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "workspaceId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "invitationId",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Invitation revoked"
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notifications/{notificationId}/read": {
      "post": {
        "summary": "Mark a notification as read",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "notificationId",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Marked read"
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/me/notification-preferences": {
      "get": {
        "summary": "Get notification preferences",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Preferences",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPreferences"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update notification preferences",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationPreferencesBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated preferences",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPreferences"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/push-devices": {
      "post": {
        "summary": "Register Expo push token",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPushDeviceBody"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Registered"
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Unregister Expo push token",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnregisterPushDeviceBody"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Unregistered"
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/lists/{listId}": {
      "get": {
        "summary": "Get a shopping list by id",
        "tags": [
          "Shopping Lists"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "listId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Shopping list DTO",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShoppingListDTO"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found or not a member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update a shopping list",
        "tags": [
          "Shopping Lists"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "listId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateShoppingListBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated shopping list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShoppingListDTO"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found or not a member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Archive a shopping list",
        "tags": [
          "Shopping Lists"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "listId",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Shopping list archived"
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found or not a member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}